Implement code to check if a remote repository is already added

+++ src/commands/reset.rs
@@ -11,6 +11,8 @@ pub enum Error {
     SetupRelated(String),
     #[error("{0}")]
     GetStorageFolderPath(String),
+    #[error("You need to set a remote repository before use DFM")]
+    SetRemoteRepository,
     #[error("Something wrong happened: {0}, when trying to: {1}")]
     Unknown(String, &'static str),
 }
@@ -37,6 +39,10 @@ impl Command for Reset {
                 ))
             }
         };
+
+        if utils::check_if_remote_link_is_added(&storage_folder_path).is_err() {
+            return Err(Error::SetRemoteRepository);
+        }
 
         if let Err(err) = fs::remove_dir_all(&storage_folder_path) {
             return Err(Error::Unknown(err.to_string(), "remove the storage folder"));
